rupees migration: peripheral columns + backend /100 sweep + Decimal JSON
- stationery (job/service) + group_payment_splits -> Numeric rupees (migration 0042)
- strip /100 in settlement/dashboard/kpi/fraud/rewards/forecast/ml/refund-eta
- Razorpay boundary to_paise() for initiate + refund
- global ENCODERS_BY_TYPE[Decimal]=float so amounts serialize as JSON numbers
- order/group event publishers use Decimal-safe json_default
- 39/40 money tests green (E2E, payments, stationery, ledger, qr-audit)

+++ tnt-backend-main/app/modules/group_cart/model.py
@@ -113,7 +113,7 @@ class GroupPaymentSplit(Base):
     group_id = Column(Integer, ForeignKey("groups.id"), nullable=False)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     split_type = Column(Enum(PaymentSplitType, values_callable=lambda x: [e.value for e in x]), default=PaymentSplitType.EQUAL)
-    amount = Column(Float, nullable=True)  # For custom splits
+    amount = Column(Numeric(10, 2), nullable=True)  # rupees, for custom splits
     percentage = Column(Float, nullable=True)  # For percentage-based splits
     created_at = Column(DateTime, default=utcnow_naive)
